jdbc driver, reorganise
