+++ pom.xml
@@ -55,11 +55,20 @@
 			<artifactId>junit-addons</artifactId>
 			<version>1.4</version>
 		</dependency> -->
-<!-- 		<dependency>
+		<!-- <dependency>
 		<groupId>com.oracle</groupId>
 		<artifactId>ojdbc14</artifactId>
 		<version>10.2.0.5</version>
 		</dependency> -->
+
+	<dependency>
+		<groupId>com.oracle</groupId>
+		<artifactId>ojdbc14</artifactId>
+		<version>10.2.0.5</version>
+		<scope>system</scope>
+		<!--<systemPath>C:/Users/prasad01/tools/jars/oracle/ojdbc14/10.2.0.5/ojdbc14-10.2.0.5.jar</systemPath> -->
+		<systemPath>/support/home/esradm/doNotDelete/ojdbc14-10.2.0.5.jar</systemPath>
+	</dependency>
 
 	</dependencies>
 </project>
