--- src/main/java/com/kfplc/ci/stock/App.java
@@ -1,13 +1,0 @@
-package com.kfplc.ci.stock;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
