+++ src/main/java/com/kfplc/ci/stock/util/ProtectedConfig.java
@@ -1,4 +1,4 @@
-package com.kfplc.ci.stock;
+package com.kfplc.ci.stock.util;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
